+++ public/data/mainData.json
@@ -5,7 +5,7 @@
       "userId": "제이지",
       "InfoTime": "1시간",
       "usrIdSrc": "images/chaebinhan/Main/1.jpg",
-      "usrFeedSrc": ["images/chaebinhan/Main/4.jpg", "images/chaebinhan/Main/6.jpg", "images/chaebinhan/Main/3.jpg"],
+      "usrFeedSrc": ["images/chaebinhan/Main/3.jpg", "images/chaebinhan/Main/5.jpg", "images/chaebinhan/Main/7.jpg"],
       "goodCount": "2312",
       "bodyTitle": "퇴근길 마음만은 파리 투나잇🎶요 디제 펑키스 파레",
       "bodyContents": "현실은 거리두기 2.5 단계ㅠㅠ오늘도 내적 댄스로 대리만족 (ﾉ∀`♥)",
@@ -24,7 +24,7 @@
       "bodyTitle": "어때요, 우리 제법 잘 어울리나요?",
       "bodyContents": "왕 큰 빅 자이언트 소파 필요한 집돌 피플 지금 당장 클릭!",
       "footerReplyCount": "122",
-      "footerReplyId": "김*인",
+      "footerReplyId": "한*빈",
       "footerReplyContent": "동묘앞 프렌즈가 최고야!",
       "feedLike": false
     },
@@ -38,7 +38,7 @@
       "bodyTitle": "올해는 집콕 파티지!",
       "bodyContents": "생각했던거보다 쬐~끔 작은게 왔네",
       "footerReplyCount": "8676",
-      "footerReplyId": "이*영",
+      "footerReplyId": "한*빈",
       "footerReplyContent": "동묘앞 프렌즈가 최고야!",
       "feedLike": false
     },
