--- src/Pages/MainDetail/DetailFeeds/DetailFeeds.scss
@@ -1,95 +1,0 @@
-@import "../../../styles/common.scss";
-
-.slick-next:before,
-.slick-prev:before {
-  color: #888;
-  font-size: 27px;
-  margin-left: -3px;
-}
-
-.feedShareBtn {
-  @include btn_mix();
-}
-.shareIcon {
-  background-color: #fff;
-  height: 27px;
-  width: 27px;
-  cursor: pointer;
-}
-
-.myModal {
-  display: none;
-  position: fixed;
-  z-index: 1;
-  padding-top: 360px;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-  background-color: rgba(0, 0, 0, 0.7);
-
-  .modal-content {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    background-color: #fefefe;
-    margin: auto;
-    padding: 20px;
-    border: 1px solid #888;
-    height: 232px;
-    width: 310px;
-    border-radius: 8px;
-
-    .modalSectionHight {
-      display: flex;
-      align-items: center;
-      flex-direction: column;
-
-      .shareText {
-        font-size: 16px;
-        font-weight: bold;
-        line-height: 1.5;
-        letter-spacing: -0.25px;
-        color: rgb(0, 0, 0);
-        // text-align: center;
-        padding-top: 10px;
-        padding-bottom: 10px;
-      }
-
-      .modalSectiondawn {
-        display: flex;
-        justify-content: center;
-
-        .shareIconItem {
-          border-radius: 70%;
-          height: 60px;
-          width: 60px;
-          overflow: hidden;
-          margin: 10px;
-          margin-bottom: 45px;
-          cursor: pointer;
-        }
-      }
-
-      .close {
-        display: flex;
-        justify-content: center;
-        border-top: 1px solid rgb(228, 227, 227);
-        width: 310px;
-        color: #aaaaaa;
-        font-size: 16px;
-        font-weight: bold;
-        line-height: 49px;
-        letter-spacing: -0.21px;
-        border-width: 1px;
-        &:hover,
-        &:focus {
-          color: #000;
-          text-decoration: none;
-          cursor: pointer;
-        }
-      }
-    }
-  }
-}
