+++ src/Routes.js
@@ -16,7 +16,7 @@ class Routes extends React.Component {
       <Router>
         <Switch>
           <Route exact path="/" component={Main} />
-          <Route exact path="/MainDetail" component={MainDetail} />
+          {/* <Route exact path="/MainDetail" component={MainDetail} /> */}
           <Route exact path="/CartList" component={CartList} />
           <Route exact path="/ProductDetail" component={ProductDetail} />
           <Route exact path="/Checkbox" component={Checkbox} />
