+++ src/styles/common.scss
@@ -1,7 +1,7 @@
 @import "reset";
 
 $theme-color: blue;
-$hr-gray: #e7e7e7;
+$bg-color: red;
 
 @mixin flex_center {
   display: flex;
